+++ MJ_FB_Backend/AGENTS.md
@@ -31,7 +31,7 @@
 - A daily reminder job (`src/utils/bookingReminderJob.ts`) emails clients about next-day bookings using the `enqueueEmail` queue. It runs nightly at 7:00 PM Regina time via `node-cron` (`0 19 * * *`) and exposes `startBookingReminderJob`/`stopBookingReminderJob`.
 - A volunteer shift reminder job (`src/utils/volunteerShiftReminderJob.ts`) emails volunteers about next-day shifts on the same schedule. It runs nightly at 7:00 PM Regina time and exposes `startVolunteerShiftReminderJob`/`stopVolunteerShiftReminderJob`.
  - A nightly no-show cleanup job (`src/utils/noShowCleanupJob.ts`) uses `node-cron` with `0 19 * * *` Regina time to mark past approved bookings as `no_show` and does not run on server startup. It exposes `startNoShowCleanupJob`/`stopNoShowCleanupJob`.
- - A nightly volunteer no-show cleanup job (`src/utils/volunteerNoShowCleanupJob.ts`) uses `node-cron` with `0 19 * * *` Regina time to mark past approved volunteer bookings as `no_show` and does not run on server startup. It logs results and waits `VOLUNTEER_NO_SHOW_HOURS` (default `24`) after each shift before auto-marking.
+ - A nightly volunteer no-show cleanup job (`src/jobs/volunteerNoShowCleanupJob.ts`) uses `node-cron` with `0 19 * * *` Regina time to mark past approved volunteer bookings as `no_show` and does not run on server startup. It logs results and waits `VOLUNTEER_NO_SHOW_HOURS` (default `24`) after each shift before auto-marking.
 - Failures in these cleanup jobs trigger alerts via Telegram when `TELEGRAM_BOT_TOKEN` and `TELEGRAM_ALERT_CHAT_ID` are configured. Alerts include a timestamp and stack trace for easier debugging.
 - API requests that result in 5xx responses also trigger Telegram alerts including the HTTP method, path, timestamp, and stack trace when these environment variables are set.
 - A nightly retention job (`src/utils/bookingRetentionJob.ts`) removes `bookings` and `volunteer_bookings` older than one year and aggregates volunteer stats into the `volunteers` table.
