+++ MJ_FB_Backend/tests/volunteerNoShowCleanupJob.test.ts
@@ -1,13 +1,6 @@
 jest.mock('node-cron', () => ({ schedule: jest.fn() }), { virtual: true });
-jest.mock('../src/utils/scheduleDailyJob', () => {
-  const actual = jest.requireActual('../src/utils/scheduleDailyJob');
-  return {
-    __esModule: true,
-    default: (cb: any, schedule: string) => actual.default(cb, schedule, false, false),
-  };
-});
 jest.mock('../src/utils/opsAlert');
-const job = require('../src/utils/volunteerNoShowCleanupJob');
+const job = require('../src/jobs/volunteerNoShowCleanupJob');
 const {
   cleanupVolunteerNoShows,
   startVolunteerNoShowCleanupJob,
